Extract SelectSystem to separate module
To keep the system module focused on the System trait, this commit
extracts the SelectSystem struct and related items to a new file.
